added Discord bridge functionality
