+++ .idea/libraries/JDA_4_2_0_168_withDependencies_min.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="JDA-4.2.0_168-withDependencies-min">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/res/lib/JDA-4.2.0_168-withDependencies-min.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
