+++ src/gg/codie/mineonline/LauncherFiles.java
@@ -23,6 +23,8 @@ public class LauncherFiles {
     public static final String ASM_JAR = MINEONLINE_LIBRARY_FOLDER + "asm-8.0.1.jar";
     public static final String ASM_COMMONS_JAR = MINEONLINE_LIBRARY_FOLDER + "asm-commons-8.0.1.jar";
     public static final String DISCORD_RPC_JAR = MINEONLINE_LIBRARY_FOLDER + "discord-rpc.jar";
+    public static final String JDA_JAR = MINEONLINE_LIBRARY_FOLDER + "JDA-4.2.0_168-withDependencies-min.jar";
+    public static final String WEBHOOK_JAR = MINEONLINE_LIBRARY_FOLDER + "discord-webhooks-0.5.0.jar";
 
     public static final String MINEONLINE_NATIVES_FOLDER = MINEONLINE_LIBRARY_FOLDER + "native" + File.separator + OSUtils.getPlatform().toString();
