+++ src/gg/codie/mineonline/LibraryManager.java
@@ -134,6 +134,8 @@ public class LibraryManager {
         classpath.append(getClasspathSeparator() + LauncherFiles.BYTEBUDDY_JAR);
         classpath.append(getClasspathSeparator() + LauncherFiles.ASM_COMMONS_JAR);
         classpath.append(getClasspathSeparator() + LauncherFiles.ASM_JAR);
+        classpath.append(getClasspathSeparator() + LauncherFiles.JDA_JAR);
+        classpath.append(getClasspathSeparator() + LauncherFiles.WEBHOOK_JAR);
 
         for(String jar : includeJars) {
             classpath.append(getClasspathSeparator() + jar);
