+++ src/gg/codie/mineonline/discord/MessageRecievedListener.java
@@ -1,0 +1,7 @@
+package gg.codie.mineonline.discord;
+
+public interface MessageRecievedListener {
+
+    void onMessageRecieved(String message);
+
+}
