+++ src/gg/codie/mineonline/server/Properties.java
@@ -6,6 +6,9 @@ public class Properties extends gg.codie.minecraft.server.Properties {
     protected static final String SERVERLIST_IP = "serverlist-ip";
     protected static final String SERVERLIST_PORT = "serverlist-port";
     protected static final String VERSION_MD5 = "version-md5";
+    protected static final String DISCORD_TOKEN = "discord-token";
+    protected static final String DISCORD_CHANNEL = "discord-channel";
+    protected static final String DISCORD_WEBHOOK_URL = "discord-webhook-url";
 
     public Properties(String jarPath) throws IOException {
         super(jarPath);
@@ -26,6 +29,12 @@ public class Properties extends gg.codie.minecraft.server.Properties {
             return super.serverPort();
         }
     }
+
+    public String discordToken() { return properties.getProperty(DISCORD_TOKEN, null); }
+
+    public String discordChan() { return properties.getProperty(DISCORD_CHANNEL, null); }
+
+    public String discordWebhookUrl() { return properties.getProperty(DISCORD_WEBHOOK_URL, null); }
 
     public String versionMD5() {
         return properties.getProperty(VERSION_MD5, null);
